♻ refactor : [common 함수 리팩토링] | [김은하] | [splash onboarding choosewho common]
