+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/common/lottie/LottieLoader.kt
@@ -1,0 +1,19 @@
+package com.shbhack.eggmoneyna.ui.common.lottie
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieCompositionSpec
+import com.airbnb.lottie.compose.LottieConstants
+import com.airbnb.lottie.compose.rememberLottieComposition
+
+@Composable
+fun LottieLoader(source: Int, modifier: Modifier) {
+    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(source))
+    LottieAnimation(
+        modifier = modifier,
+        composition = composition,
+        iterations = LottieConstants.IterateForever
+    )
+}
