+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/common/system/SetSsytemBarColor.kt
@@ -1,0 +1,20 @@
+package com.shbhack.eggmoneyna.ui.common.system
+
+import android.app.Activity
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalView
+
+@Composable
+fun SetSystemBarColor(color: Color) {
+    val view = LocalView.current
+    if (!view.isInEditMode) {
+        SideEffect {
+            val window = (view.context as Activity).window
+            window.statusBarColor = color.toArgb()
+            window.navigationBarColor = color.toArgb()
+        }
+    }
+}
