+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/onboarding/OnBoarding2Screen.kt
@@ -1,7 +1,6 @@
 package com.shbhack.eggmoneyna.ui.onboarding
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
@@ -24,14 +23,16 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.shbhack.eggmoneyna.R
 import com.shbhack.eggmoneyna.ui.EggMoneynaDestination
+import com.shbhack.eggmoneyna.ui.common.button.ButtonRadius40
+import com.shbhack.eggmoneyna.ui.common.lottie.LottieLoader
+import com.shbhack.eggmoneyna.ui.common.system.SetSystemBarColor
 import com.shbhack.eggmoneyna.ui.theme.contextTextColor
 import com.shbhack.eggmoneyna.ui.theme.onboardingColor2
-import com.shbhack.eggmoneyna.util.CommonUtils
 
 @Composable
 fun OnBoarding2Screen(navController: NavController) {
 
-    CommonUtils.setSystemBarColor(color = onboardingColor2)
+    SetSystemBarColor(color = onboardingColor2)
 
     Column(
         modifier = Modifier
@@ -73,13 +74,13 @@ fun OnBoarding2Screen(navController: NavController) {
                 .height(0.dp)
                 .weight(1f)
         )
-        CommonUtils.LottieLoader(
+        LottieLoader(
             source = R.raw.onboarding_compliment,
             modifier = Modifier
                 .fillMaxWidth()
                 .aspectRatio(1f)
         )
-        CommonUtils.buttonRadius40(text = stringResource(id = R.string.next), color = Color.Black) {
+        ButtonRadius40(text = stringResource(id = R.string.next), color = Color.Black) {
             navController.navigate(EggMoneynaDestination.ON_BOARDING3)
         }
     }
