+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/onboarding/OnBoarding3Screen.kt
@@ -24,13 +24,15 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.shbhack.eggmoneyna.R
 import com.shbhack.eggmoneyna.ui.EggMoneynaDestination
+import com.shbhack.eggmoneyna.ui.common.button.ButtonRadius40
+import com.shbhack.eggmoneyna.ui.common.lottie.LottieLoader
+import com.shbhack.eggmoneyna.ui.common.system.SetSystemBarColor
 import com.shbhack.eggmoneyna.ui.theme.contextTextColor
 import com.shbhack.eggmoneyna.ui.theme.onboardingColor3
-import com.shbhack.eggmoneyna.util.CommonUtils
 
 @Composable
 fun OnBoarding3Screen(navController: NavController) {
-    CommonUtils.setSystemBarColor(color = onboardingColor3)
+    SetSystemBarColor(color = onboardingColor3)
 
     Column(
         modifier = Modifier
@@ -51,7 +53,7 @@ fun OnBoarding3Screen(navController: NavController) {
                 navController.navigate(EggMoneynaDestination.CHOOSE_WHO)
             },
         )
-        CommonUtils.LottieLoader(
+        LottieLoader(
             source = R.raw.onboarding_egg,
             modifier = Modifier
                 .fillMaxWidth()
@@ -83,7 +85,7 @@ fun OnBoarding3Screen(navController: NavController) {
                 .weight(1f)
         )
 
-        CommonUtils.buttonRadius40(text = stringResource(id = R.string.next), color = Color.Black) {
+        ButtonRadius40(text = stringResource(id = R.string.next), color = Color.Black) {
             navController.navigate(EggMoneynaDestination.ON_BOARDING4)
         }
     }
