+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/onboarding/OnBoarding5Screen.kt
@@ -19,14 +19,16 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.shbhack.eggmoneyna.R
 import com.shbhack.eggmoneyna.ui.EggMoneynaDestination
+import com.shbhack.eggmoneyna.ui.common.button.ButtonRadius40
+import com.shbhack.eggmoneyna.ui.common.lottie.LottieLoader
+import com.shbhack.eggmoneyna.ui.common.system.SetSystemBarColor
 import com.shbhack.eggmoneyna.ui.theme.contextTextColor
 import com.shbhack.eggmoneyna.ui.theme.onboardingColor5
-import com.shbhack.eggmoneyna.util.CommonUtils
 
 @Composable
 fun OnBoarding5Screen(navController: NavController) {
 
-    CommonUtils.setSystemBarColor(color = onboardingColor5)
+    SetSystemBarColor(color = onboardingColor5)
 
     Column(
         modifier = Modifier
@@ -56,13 +58,13 @@ fun OnBoarding5Screen(navController: NavController) {
                 .height(0.dp)
                 .weight(1f)
         )
-        CommonUtils.LottieLoader(
+        LottieLoader(
             source = R.raw.onboarding_plan,
             modifier = Modifier
                 .fillMaxWidth()
                 .aspectRatio(1f)
         )
-        CommonUtils.buttonRadius40(text = stringResource(id = R.string.start), color = Color.Black) {
+        ButtonRadius40(text = stringResource(id = R.string.start), color = Color.Black) {
             navController.navigate(EggMoneynaDestination.CHOOSE_WHO)
         }
     }
