+++ android/app/src/main/java/com/shbhack/eggmoneyna/ui/onboarding/OnBoardingScreen.kt
@@ -23,14 +23,16 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.shbhack.eggmoneyna.R
 import com.shbhack.eggmoneyna.ui.EggMoneynaDestination
+import com.shbhack.eggmoneyna.ui.common.button.ButtonRadius40
+import com.shbhack.eggmoneyna.ui.common.lottie.LottieLoader
+import com.shbhack.eggmoneyna.ui.common.system.SetSystemBarColor
 import com.shbhack.eggmoneyna.ui.theme.contextTextColor
 import com.shbhack.eggmoneyna.ui.theme.onboardingColor1
-import com.shbhack.eggmoneyna.util.CommonUtils
 
 @Composable
 fun OnBoardingScreen(navController: NavController) {
 
-    CommonUtils.setSystemBarColor(color = onboardingColor1)
+    SetSystemBarColor(color = onboardingColor1)
 
     Column(
         modifier = Modifier
@@ -72,14 +74,14 @@ fun OnBoardingScreen(navController: NavController) {
                 .height(0.dp)
                 .weight(1f)
         )
-        CommonUtils.LottieLoader(
+        LottieLoader(
             source = R.raw.onboarding_summary,
             modifier = Modifier
                 .fillMaxWidth()
                 .aspectRatio(1f)
                 .padding(bottom = 20.dp)
         )
-        CommonUtils.buttonRadius40(text = stringResource(id = R.string.next), color = Color.Black) {
+        ButtonRadius40(text = stringResource(id = R.string.next), color = Color.Black) {
             navController.navigate(EggMoneynaDestination.ON_BOARDING2)
         }
     }
